refactor: update type decimal to float in payments module

+++ src/payments/interfaces.py
@@ -1,6 +1,5 @@
 from abc import ABC, abstractmethod
 from typing import Dict, Any, Optional, List
-from decimal import Decimal
 
 from database.models.orders import OrderModel
 from database.models.payments import PaymentModel, PaymentStatusEnum
@@ -11,7 +10,7 @@ class PaymentServiceInterface(ABC):
     async def create_payment_intent(
         self,
         order: OrderModel,
-        amount: Decimal,
+        amount: float,
         currency: str = "usd"
     ) -> Dict[str, Any]:
         pass
@@ -37,7 +36,7 @@ class PaymentServiceInterface(ABC):
     async def process_refund(
         self,
         payment: PaymentModel,
-        amount: Optional[Decimal] = None,
+        amount: Optional[float] = None,
         reason: Optional[str] = None
     ) -> Dict[str, Any]:
         pass
